Fix warnings
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@796 3c1deb5b-d424-0410-962d-aba41a686d42

+++ source/unemap/beekeeper.c
@@ -1699,7 +1699,7 @@ if (!found)
 	long temp_long;
 	unsigned short temp_short;
 
-	printf("time %ld\n",time);
+	printf("time %ld\n",(long int)(time));
 	for (i=0;i<number_of_channels;i++)
 	{
 		if (0xFFFF!=gain[i])
